+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 **checkout**
 * add config `commerce.checkout.orderService.skipCartValidation` to disable cart validation that happens right before placing an order in the OrderService
 
+**search**
+* **Breaking:** fixed fake service to interpret sort options and gives away empty promotion
+* removed deprecation notice from fields `Asc` and `Desc` in `SortOption` struct
+
 ## v3.11.0
 
 **search**